+++ main/src/main/java/net/oneandone/stool/configuration/StoolConfiguration.java
@@ -68,11 +68,6 @@ public class StoolConfiguration {
     public String admin;
 
     /**
-     * true if multiple users work on stages
-     */
-    public boolean shared;
-
-    /**
      * true if users have to commit all source changes before stool allows them to start the stage.
      */
     public boolean committed;
@@ -127,7 +122,6 @@ public class StoolConfiguration {
         diskMin = 1000;
         admin = "";
         autoRemove = -1;
-        shared = false;
         committed = false;
         defaults = new LinkedHashMap<>();
         defaults.put("", new HashMap<>());