+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 5.0.0 (pending)
 
 * removed `sleep` state and the corresponding `stop -sleep` flag, it was never used
+* removed `shared` switch - local maven repository is now always the user's Maven repository
 * implementation changes
   * split Stage Class into a Project- and a new Stage class: project is everything around the former stage directory (which is 
     typically the checkout); the new Stage roughly represents the backstage directory